% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{read_data}
\alias{read_data}
\title{Read UCI or Physionet 2012 Challenge Data}
\usage{
read_data(dataset, sim_params, sim_index)
}
\arguments{
\item{dataset}{String for name of dataset. Valid datasets: "BANKNOTE","...."}

\item{sim_params}{List of simulation parameters (N, D, P, seed)}

\item{sim_index}{Index of simulation run (for setting seeds)}
}
\value{
list of objects: data (N x P matrix), classes (subgroups of observations), and params (those used for simulating data)
}
\description{
Read UCI or Physionet 2012 Challenge Data
}
\examples{
read_data("BANKNOTE",NULL,1)
}
