% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(
  N,
  D,
  P,
  sim_index,
  seed = 9 * sim_index,
  ratio = c(6, 2, 2),
  g_seed = 333
)
}
\arguments{
\item{ratio}{Train-valid-test ratio for splitting of observations}

\item{g_seed}{Seed for train-valid-test dataset splitting of observations}
}
\value{
list of objects: data (N x P matrix), classes (subgroups of observations), params (those used for simulating data), and g (partitioning of data into train-valid-test sets)
}
\description{
Simulate data
}
\examples{
simulate_data(N = 10000, D = 2, P = 8, sim_index = 1)
}
