% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneHyperparams.R
\name{tuneHyperparams}
\alias{tuneHyperparams}
\title{Hyperparameter tuning used by NIMIWAE() function}
\usage{
tuneHyperparams(
  method = "NIMIWAE",
  data,
  Missing,
  g,
  rdeponz = F,
  learn_r = T,
  phi0 = NULL,
  phi = NULL,
  ignorable = F,
  covars_r = rep(1, ncol(data)),
  arch = "IWAE",
  sigma = "elu",
  h = c(128L, 64L),
  n_hidden_layers = c(1L, 2L),
  n_hidden_layers_r = 0L,
  bs = 1000L,
  lr = c(0.001, 0.01),
  dim_z = as.integer(c(floor(p/2), floor(p/4))),
  niws = 5L,
  n_epochs = 2002L,
  data_types = NULL,
  one_hot_max_sizes = NULL,
  ohms = NULL,
  MissingDatas = NULL
)
}
\arguments{
\item{method}{String specifying the method to tune hyperparameters. Can be "NIMIWAE" (default), "MIWAE", "VAEAC", "HIVAE", "MEAN", or "MF". Hyperparameters are not tuned for "MEAN" or "MF"}

\item{data}{Data matrix (N x P)}

\item{Missing}{Missingness mask matrix (N x P)}

\item{g}{Training-validation-test split partitioning}

\item{rdeponz}{TRUE/FALSE: Whether to allow missingness (r) to depend on the latent variable (z). Default is FALSE}

\item{learn_r}{TRUE/FALSE: Whether to learn missingness model via appended NN (TRUE, default), or fit a known logistic regression model (FALSE). If FALSE, `phi0` and `phi` must be specified}

\item{phi0}{(optional) Intercept of logistic regression model, if learn_r = FALSE.}

\item{phi}{(optional) Vector of coefficients of logistic regression model for each input covariates `covars_r`, if learn_r = FALSE. `phi` must be the same length as the number of input covariates, or `sum(covars_r)`.}

\item{ignorable}{TRUE/FALSE: Whether missingness is ignorable (MCAR/MAR) or nonignorable (MNAR, default). If missingness is known to be ignorable, "ignorable=T" omits missingness model.}

\item{covars_r}{Vector of 1's and 0's of whether each feature is included as covariates in the missingness model. Need not be specified if `ignorable = T`. Default is using all features as covariates in missingness model. Must be length P (or `ncol(data)`)}

\item{arch}{Architecture of NIMIWAE. Can be "IWAE" or "VAE". "VAE" is specific case of the "IWAE" where only one sample is drawn from the joint posterior of (z, xm).}

\item{sigma}{activation function ("relu" or "elu")}

\item{h}{integer, number of nodes per hidden layer}

\item{n_hidden_layers}{integer, #hidden layers (except missingness model Decoder_r)}

\item{n_hidden_layers_r}{integer, #hidden layers for Decoder_r (default: 0)}

\item{bs}{integer, batch size (training)}

\item{lr}{float, learning rate}

\item{dim_z}{integer, dimensionality of latent z. Default: 5}

\item{data_types}{Specify for HIVAE only.}

\item{one_hot_max_sizes}{Specify for VAEAC only.}

\item{ohms}{Specify for VAEAC only.}

\item{MissingDatas}{Specify for VAEAC only.}

\item{niw}{integer, number of importance weights (samples drawn from each latent space). Default: 5}

\item{n_epoch}{integer, maximum number of epochs (without early stop). Default: 2002}
}
\value{
res object: method's fit on test set, after training on training set and validating best set of hyperparameter values using the validation set.
}
\description{
Hyperparameter tuning used by NIMIWAE() function
}
