% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(
  N,
  D,
  P,
  sim_index,
  seed = 9 * sim_index,
  ratio = c(6, 2, 2),
  g_seed = 333
)
}
\arguments{
\item{N}{Number of observations}

\item{D}{Dimension of latent Z}

\item{P}{Number of features}

\item{sim_index}{Integer index for sim. Can use to set seed}

\item{seed}{Seed. Either set custom or set some value x sim_index. Default: 9 x sim_index}

\item{ratio}{Train-valid-test ratio for splitting of observations}

\item{g_seed}{Seed for train-valid-test dataset splitting of observations}
}
\value{
list of objects: data (N x P matrix), classes (subgroups of observations), params (those used for simulating data), and g (partitioning of data into train-valid-test sets)
}
\description{
Simulate data
}
\examples{
simulate_data(N = 10000, D = 2, P = 8, sim_index = 1)

}
\references{
\url{https://github.com/DavidKLim/NIMIWAE}
}
\author{
David K. Lim, \email{deelim@live.unc.edu}
}
